[FEAT] Finish threading pub_arm; pub_arm_with_speed is not yet applied
